Fix useData() hooks and add testing.
The existing useData() hooks didn't gracefully handle passing the same metrics / regions on every
render if the actual array instances changed, e.g.: This would cause an infinite re-render:

catalog.useDataForRegionsAndMetrics([region], [metric])

Because the array instances would be differente every time.  The core fix is to add a
`useCachedArrayIfEqual()` helper that preserves the original array reference if the contents are the
same.