Fix CovidActNowDataProvider caching (#241) and rework caching in general.
It turns out that CachingMetricDataProviderBase wasn't that helpful for implementing
a MetricDataProvider with caching. It kind of works for a provider that just has one big asset to
cache (like the CsvDataProvider, though we messed it up on the first try). But it does not work well
when you have more granular assets to cache (e.g. one per region like with CAN).  So I have:

* Replaced CachingMetricDataProviderBase with SimpleMetricDataProviderBase that helps you write a
  provider that fetches a single region and metric at a time but doesn't pretend to help with
  caching.
* Reworked the CovidActNowDataProvider to just implement MetricDataProvider directly and correctly
  cache data on a per-region basis, and avoid having multiple outstanding requests for the same
  region file.
* Minor cleanup of CsvDataProvider to simplify the caching implementation.

Fixes #241.